DOOM-ML-PROJECT/DatasetGenerator/requirements.txt
matplotlib==3.2.1
numpy==1.17.4
pandas==0.25.3
Pillow==7.1.2
